Break gui.py and tools.py into new files
