+++ gui/__init__.py
@@ -1,3 +1,1 @@
 from .gui import MainDialog
-from .config import ConfigDialog
-from .tools import StatusThread, ParseThread, UploadThread, QtHandler
